requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas>=2.0.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.0
openpyxl>=3.1.0
reportlab>=4.0.0
pydantic>=2.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0